pyproject.toml
[project]
name = "mcp-oauth-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp==2.13.3",
    "python-dotenv>=1.1.1",
    "pydantic>=2.12.3"
]
